Overhaul to work better with Razor Pages
Previously we've relied on having a controller to group all actions
within a journey. With this, we have a single place to describe the
journey and this was done with the JourneyMetadataAttribute. This both
described the journey itself and annotated the actions within the
controller so that we know what journey the actions belong to.

For Razor Pages, this doesn't work; there's no single place to annotate
all pages within a given journey so we need another way to register
metadata for the journey. This is now done via JourneyRegistry on
FormFlowOptions. For now, each journey must be explicitly registered.
There's a new JourneyAttribute which takes a journey name that indicates
that a controller, action, page or page handler belongs to a specific
journey. The JourneyMetadataAttribute is no more.

There are a few tweaks to various other components too. Notably, the
methods on JourneyInstanceProvider now take an ActionContext parameter
instead of the implementations resolving ActionContext itself.